requirements.txt
Django==5.2.6
djangorestframework==3.15.2
psycopg[binary]>=3.1,<4
python-dotenv==1.0.1

# OpenAI + LangChain
openai>=1.40.0
langchain>=0.2.12
langchain-community>=0.2.10
langchain-openai>=0.2.0
tiktoken>=0.7.0

# Utilidades
Faker>=25.0.0
django-cors-headers>=4.4.0
gunicorn>=22.0.0


drf-spectacular>=0.27
# Opcional (UI sin internet/CDN):
drf-spectacular[sidecar]>=0.27
djangorestframework>=3.15
drf-spectacular>=0.27.2
drf-spectacular-sidecar>=2024.7.1
